+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2024-12-28
+- [Add ComfyUI_LayerStyle](https://github.com/chflame163/ComfyUI_LayerStyle) custom node
+
+- Add ComfyUI/models/checkpoints/DreamShaper_XL_v2_1_TurboDPMDER.safetensors to checkpoints
 ## 2024-12-27
 - [Add ComfyUI-LogicUtils](https://github.com/aria1th/ComfyUI-LogicUtils) custom node
 - [Add ComfyUI-WD14-Tagger](https://github.com/pythongosssss/ComfyUI-WD14-Tagger) custom node